code/Backend/requirements.txt
flask
faiss-cpu
torch
sentence-transformers
torchvision
Pillow
tqdm
